trying to animate the shopping list
