+++ src/App.jsx
@@ -6,6 +6,7 @@ import { createGlobalStyle } from 'styled-components';
 import { QueryClient, QueryClientProvider } from '@tanstack/react-query';
 import { ReactQueryDevtools } from '@tanstack/react-query-devtools';
 import { SidePageProvider } from './Context/SidePageProvider';
+import AppProvider from './Context/AppContext';
 
 /*
 const router = createBrowserRouter([
@@ -99,14 +100,16 @@ function App() {
   return (
     <QueryClientProvider client={queryClient}>
       <ReactQueryDevtools />
-      <SidePageProvider>
-        <BrowserRouter>
-          <GlobalStyle />
-          <AnimatePresence>
-            <RoutesWithAnimation />
-          </AnimatePresence>
-        </BrowserRouter>
-      </SidePageProvider>
+      <AppProvider>
+        <SidePageProvider>
+          <BrowserRouter>
+            <GlobalStyle />
+            <AnimatePresence>
+              <RoutesWithAnimation />
+            </AnimatePresence>
+          </BrowserRouter>
+        </SidePageProvider>
+      </AppProvider>
     </QueryClientProvider>
   );
 }
