+++ src/Variables/variables.js
@@ -29,3 +29,18 @@ export const childrenVariants = {
     },
   },
 };
+
+export const listChildrenVariants = {
+  initial: {
+    opacity: 0,
+    x: '30px',
+  },
+  final: {
+    opacity: 1,
+    x: '0px',
+    transition: {
+      duration: 0.6,
+      delay: 0.4,
+    },
+  },
+};
